Optimizations

+++ src/raytrace-rs/src/configuration.rs
@@ -5,26 +5,26 @@ use serde::{Deserialize, Serialize};
 #[derive(Serialize, Deserialize, Debug)]
 pub struct RaytracerScene {
     pub multithreading: bool,
-    pub aspect_ratio: f64,
+    pub aspect_ratio: f32,
     pub image_width: i32,
     pub image_height: i32,
     pub samples_per_pixel: usize,
     pub max_depth: i32,
-    pub v_fov: f64,
-    pub aperture: f64,
-    pub focal_distance: f64,
-    pub camera_pos: Vec<f64>,
-    pub camera_dir: Vec<f64>,
-    pub camera_up: Vec<f64>,
+    pub v_fov: f32,
+    pub aperture: f32,
+    pub focal_distance: f32,
+    pub camera_pos: Vec<f32>,
+    pub camera_dir: Vec<f32>,
+    pub camera_up: Vec<f32>,
     pub objects: Vec<RonObject>,
-    pub lights: Vec<Vec<f64>>,
+    pub lights: Vec<Vec<f32>>,
 }
 
 #[derive(Clone, Serialize, Deserialize, Debug)]
 pub struct RonObject {
     pub objtype: String,
-    pub vectors: Vec<Vec<f64>>,
-    pub scalars: Vec<f64>,
+    pub vectors: Vec<Vec<f32>>,
+    pub scalars: Vec<f32>,
     pub material: Vec<String>,
 }
